requirements.txt
torch
torchvision
pillow
kornia
transformers
gradio
timm